# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de LearnUp",
        "description": "Documentación de la API de LearnUp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/courses/{curso}/inscripcion": {
            "post": {
                "tags": [
                    "curso"
                ],
                "summary": "Inscribe user in a course",
                "description": "Inscribe the authenticated user in the specified course.",
                "operationId": "inscribeUserInCourse",
                "parameters": [
                    {
                        "name": "curso",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Inscription successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inscripción exitosa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already enrolled in the course"
                    },
                    "402": {
                        "description": "Payment required"
                    }
                }
            }
        },
        "/courses/{curso}/students": {
            "get": {
                "tags": [
                    "curso"
                ],
                "summary": "Get all students in a course",
                "description": "Retrieve a list of all students enrolled in the specified course.",
                "operationId": "getAllStudentsInCourse",
                "parameters": [
                    {
                        "name": "curso",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filter students by name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "john@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/images": {
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "Store image",
                "description": "Upload an image to S3 and update the database.",
                "operationId": "storeImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file",
                                    "table",
                                    "row_id",
                                    "column"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "table": {
                                        "type": "string",
                                        "example": "users"
                                    },
                                    "row_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "column": {
                                        "type": "string",
                                        "example": "profile_picture"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploaded successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "images"
                ],
                "summary": "Delete image",
                "description": "Delete an image from S3 and update the database.",
                "operationId": "deleteImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table",
                                    "row_id",
                                    "column"
                                ],
                                "properties": {
                                    "table": {
                                        "type": "string",
                                        "example": "users"
                                    },
                                    "row_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "column": {
                                        "type": "string",
                                        "example": "profile_picture"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image deleted successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}